feat(plugins): support cache plugin
