+++ deploy/plugins/kubegems-local-stack/values.yaml
@@ -1,3 +1,4 @@
+imageRegistry: ""
 clusterName: local
 storageClass: local-path
 logging:
